Parse PG dates w/ 6-digit milliseconds, throw on error
Modern PG returns dates w/ more granularity, e.g.

  '2024-11-25 16:01:36.549802+00'

This format is now detected, note that millisecs
haven't been parsed before and are still not.

This was replaced w/ "now" in the response ... Never
do that :-) It now raises an exception.

Also raise exception in SQLite.
MySQL doesn't use this anymore, probably directly
reading the timestamp somehow, added a comment on
that.